+++ src/lib/index.js
@@ -4,6 +4,7 @@ import Circos from './components/Circos.react';
 import FornaContainer from './components/FornaContainer.react';
 import Ideogram from './components/Ideogram.react';
 import Igv from './components/Igv.react';
+import Pileup from './components/Pileup.react';
 import Molecule2dViewer from './components/Molecule2dViewer.react';
 import Molecule3dViewer from './components/Molecule3dViewer';
 import NeedlePlot from './components/NeedlePlot.react';
@@ -18,6 +19,7 @@ export {
     FornaContainer,
     Ideogram,
     Igv,
+    Pileup,
     Molecule2dViewer,
     Molecule3dViewer,
     NeedlePlot,